use file-scoped namespace in request models, move PageMetadataEntity to Models/Response/Pages and rename it to PageMedatata
