+++ Apps.Webflow/Models/Response/Components/ListComponentsResponse.cs
@@ -1,15 +1,14 @@
 using Blackbird.Applications.Sdk.Common;
 using Newtonsoft.Json;
 
-namespace Apps.Webflow.Models.Response.Components
-{
-    public class ListComponentsResponse
-    {
-        [Display("Components")]
-        [JsonProperty("components")]
-        public IEnumerable<ComponentResponse> Components { get; set; } = [];
+namespace Apps.Webflow.Models.Response.Components;
 
-        [JsonProperty("pagination")]
-        public PaginationInfo Pagination { get; set; } = new();
-    }
+public class ListComponentsResponse
+{
+    [Display("Components")]
+    [JsonProperty("components")]
+    public IEnumerable<ComponentResponse> Components { get; set; } = [];
+
+    [JsonProperty("pagination")]
+    public PaginationInfo Pagination { get; set; } = new();
 }
