+++ Apps.Webflow/Models/Response/Pages/UpdatePageContentResponse.cs
@@ -1,12 +1,11 @@
 ﻿using Blackbird.Applications.Sdk.Common;
 
-namespace Apps.Webflow.Models.Response.Pages
-{
-    public class UpdatePageContentResponse
-    {
-        [Display("Success")]
-        public bool Success { get; set; }
+namespace Apps.Webflow.Models.Response.Pages;
 
-        public string? Error { get; set; }
-    }
+public class UpdatePageContentResponse
+{
+    [Display("Success")]
+    public bool Success { get; set; }
+
+    public string? Error { get; set; }
 }
